Move Messages dev definitions to locales directory.
