+++ src/js/App/ErrorComponents/DefaultErrorComponent.tsx
@@ -17,7 +17,7 @@ import {
 import ExclamationCircleIcon from '@patternfly/react-icons/dist/js/icons/exclamation-circle-icon';
 import { getUrl } from '../../../utils/common';
 import { useIntl } from 'react-intl';
-import messages from '../../Messages';
+import messages from '../../../locales/Messages';
 
 import './ErrorComponent.scss';
