+++ src/js/App/GlobalFilter/TagsModal.tsx
@@ -6,8 +6,8 @@ import { fetchAllSIDs, fetchAllTags } from '../../../redux/actions';
 import debounce from 'lodash/debounce';
 import flatMap from 'lodash/flatMap';
 import { useIntl } from 'react-intl';
-import messages from '../../Messages';
 import { CommonSelectedTag, CommonTag, GlobalFilterTag, ReduxState, SID } from '../../../redux/store';
+import messages from '../../../locales/Messages';
 import { FlagTagsFilter } from './constants';
 import { TagFilterOptions, TagPagination } from './tagsApi';
 import { Action } from 'redux';
