+++ src/js/App/Header/Login.js
@@ -3,7 +3,7 @@ import { Button } from '@patternfly/react-core';
 import { login } from '../../../jwt/jwt';
 import { PageHeaderTools } from '@patternfly/react-core';
 import { useIntl } from 'react-intl';
-import messages from '../../Messages';
+import messages from '../../../locales/Messages';
 
 const Login = () => {
   const intl = useIntl();
