+++ src/js/App/QuickStart/LazyQuickStartCatalog.js
@@ -1,7 +1,7 @@
 import React, { Suspense, lazy } from 'react';
 import { LoadingBox } from '@patternfly/quickstarts';
 import { useIntl } from 'react-intl';
-import messages from '../../Messages';
+import messages from '../../../locales/Messages';
 
 const QuickStartCatalog = lazy(() => import(/* webpackChunkName: "quick-start" */ './QuickStartCatalog'));
