+++ src/layouts/DefaultLayout.js
@@ -15,7 +15,7 @@ import useOuiaTags from '../utils/useOuiaTags';
 import RedirectBanner from '../js/App/Stratosphere/RedirectBanner';
 import BarsIcon from '@patternfly/react-icons/dist/js/icons/bars-icon';
 import { useIntl } from 'react-intl';
-import messages from '../js/Messages';
+import messages from '../locales/Messages';
 import { CROSS_ACCESS_ACCOUNT_NUMBER } from '../js/consts';
 import { getUrl } from '../utils/common';
